Expose SCSI disk information as attributes
This adds the 'bus.scsi' hierarchy to the Chef attributes namespace. It
exposes information about each SCSI disk found on the node.

For each disk, the following information is exposed:
 - SCSI address: host, channel, target, lun
 - Vendor information: model, firmware version and serial number
 - Disk size and World wide number

The information is retrieved using WMI on Windows and Udev on Linux.
There is no good Udev binding in Ruby, so this implements a minimal one
using the FFI gem.