Add RHEL 6-like compatibility
When matching block device to their parent scsi_device, we were relying
on the udev_enumerate_add_match_parent method. It was preventing this
cookbook from being compatible with RHEL 6-like operating systems as
their libudev was too old.

Instead, this implements the small logic needed to do the parent
matching in the cookbook itself.