Expose PCI location for each device's host adapter
The host adapter number in the host:channel:target:lun SCSI address
depends on the enumeration order by the operating system. This makes it
difficult to precisely match a SCSI device to a physical location.

While PCI bus, device and function number are not physical identifiers,
they do not change much across the lifetime of a machine as long as the
hardware doesn't change. So, for each device, this exposes their host
adapter PCI location. This gives a somewhat stable way to identify
a device.