Mock SCSI info in RSpec when running outside this cookbook
Previously, the value Chef::Config['bus_scsi_disabled'] was set when
running in an RSpec context. This was supposed to mock/disable the
setting of the 'bus-scsi' attributes to make unit test for recipes work.

But this was subject to load ordering issue and didn't really work in
the end (especially when this cookbook was a dependency in another
cookbook).

Let's rework the system with a real mock when run from another cookbook
and no mock when run here.

While we are at it, remove support for CentOS 6.