Require 'wmi-lite' only when really needed
All Ruby files in libraries/ are loaded during Chef's compile phase. It
means that we require 'wmi-lite', an almost Windows only gem, even on
other platforms. This is not needed and may even be harmful in some
situations.

Let's play nicer and defer the inclusion to the function body where it
is really needed.