Make operations on Win32_DiskDrive properties safer
In some case, a disk doesn't have all its information exposed (this
probable happens when running in a virtual machines). This is an issue
since we try to use the .strip() method on some of those properties.

Let's use the safe navigation operator to ensure this won't trigger an
exception.